flavors update readme
