+++ README.md
@@ -1,13 +1,12 @@
 # ✅ ToDo Yandex
-Домашнее задание 3 в Школе мобильной разработки | Flutter
+Домашнее задание 4 в Школе мобильной разработки | Flutter
 
 ## ⭕️ For Reviewers
 - Огромная просьба помочь с выявлением и решением багов, улучшением кода и объяснения, как и почему лучше сделать по-другому
 - Буду очень благодарен всем комментариям, замечаниям и придиркам, даже если это стоит мне баллов
 - Также при запуске на эмуляторе релизной версии нужно использовать эмулятор x86_64 ОС
 - Мой апк сбилжен с моим токеном, для запуска в дебаге или билда релизной версии используйте --dart-define=url=/url/ --dart-define=token=/token/
-- не работает кнопка удаления на экране изменения таски
-- не доразобрался с тестами(не успел написать unit, с интеграционным что-то непонятное) 
+- По поводу доступа к Firebase обращайтесь в телеграм
 
 ## 📱 Скриншоты
 
@@ -17,16 +16,21 @@
 - Режим добавления/редактирования задачи
 - Свайп по таску выполнено/удалить
 - Показ/скрытие выполненных дел
+- Поддержка тёмной темы
+- Поддержка двух языков
 - Сохранение в локальную базу данных Isar
 - Синхронизация с сервером
 - Адекватная работа мерджа и ревизий
-- реализован Navigator 2.0 но не интегрирован
+- реализован Navigator 2.0
 - Deeplinks, вроде, работают на Android, прошу обратну связь по ним. Пример команды для запуска: (.\adb shell am start -W -a android.intent.action.VIEW -d myapp://example.com/adding)
+- Отправляется аналитика, крэшлитика в Firebase
+- Настроены Remote configs
+- Поддержка Landscape и больших экранов
 
 
 ## 📥 .apk файл для скачивания на Android
 
-Вы можете скачать приложение по ссылке: [install To Do](https://github.com/o4irov/to_do/releases/download/3.0/app-release.apk)
+Вы можете скачать приложение по ссылке: [install To Do](https://github.com/o4irov/to_do/releases/download/4.0/ToDo_release.apk)
 
 ## 📝 Библиотеки
 
@@ -44,6 +48,14 @@
 - [integration_test: sdk: flutter]
 - [shared_preferences: ^2.2.0](https://pub.dev/packages/shared_preferences)
 - [uuid: ^3.0.7](https://pub.dev/packages/uuid)
+- [firebase_core: ^2.15.0](https://pub.dev/packages/firebase_core)
+- [firebase_crashlytics: ^3.3.4](https://pub.dev/packages/firebase_crashlytics)
+- [firebase_analytics: ^10.4.4](https://pub.dev/packages/firebase_analytics)
+- [firebase_remote_config: ^4.2.3](https://pub.dev/packages/firebase_remote_config)
+- [freezed_annotation: ^2.4.1](https://pub.dev/packages/freezed_annotation)
+- [json_annotation: ^4.8.1](https://pub.dev/packages/json_annotation)
+- [get_it: ^7.6.0](https://pub.dev/packages/get_it)
+- [flutter_flavor: ^3.1.1](https://pub.dev/packages/flutter_flavor)
 
 ## 👨‍💻 Авторы
