--- launch.json
@@ -1,4 +1,0 @@
-{
-    "url": "https://beta.mrdekk.ru/todobackend",
-    "token": "cotestability" 
-}
