+++ lib/utils/logger.dart
@@ -1,4 +1,5 @@
 import 'package:logger/logger.dart';
+
 Logger get logger => Log.instance;
 
 class Log extends Logger {
