+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
   get_it: ^7.6.0
+  flutter_flavor: ^3.1.1
 
 dev_dependencies:
   flutter_test:
